[FE: feat] 게임 대기화면 및 게임 설정 수정 다이얼로그 구현
- api 연결 및 viewmodel 로직 추가

+++ fe/D104/app/src/main/java/com/d104/pnt/data/remote/model/response/GameRoomResponse.kt
@@ -5,8 +5,10 @@ import com.google.gson.annotations.SerializedName
 data class CreateGameRoomResponse(
     @SerializedName("roomId")
     val roomId: Long,
+
     @SerializedName("roomCode")
     val roomCode: String,
+
     @SerializedName("status")
     val status: String,
 )
@@ -37,10 +39,10 @@ data class GameRoomSettingsResponse(
     val cctvInterval: Int,
 
     @SerializedName("prisonLat")
-    val prisonLat: Int,
+    val prisonLat: Double,
 
     @SerializedName("prisonLng")
-    val prisonLng: Int,
+    val prisonLng: Double,
 )
 
 data class GameMemberListResponse(
